fix: 리뷰컨트롤러에 평점조회 메서드 리턴타입 변경

+++ src/main/java/com/moayo/moayoeats/backend/domain/review/controller/ReviewController.java
@@ -55,7 +55,7 @@ public class ReviewController {
     }
 
     @GetMapping("/score")
-    public ApiResponse<Integer> getScore(
+    public ApiResponse<Double> getScore(
         @AuthenticationPrincipal UserDetailsImpl userDetails
     ) {
